Adds the class to the calendar table cells
Adds the class 'calendar-cell' to each <td> tag of the calendar table layout so that each cell can be selected and manipulated by the javaScript code. This edition does add the paragraph tag that is supposed to display the actual month, second, adds the javaScript module reference to the html document.